+++ apps/backend/core/src/main/java/com/schemafy/core/project/domain/Project.java
@@ -16,14 +16,14 @@ import lombok.NoArgsConstructor;
 public class Project extends BaseEntity {
 
   private String workspaceId;
-  private Long dbVendorId;
+  private Integer dbVendorId;
   private String name;
   private String description;
 
   public static Project create(
       String id,
       String workspaceId,
-      Long dbVendorId,
+      Integer dbVendorId,
       String name,
       String description) {
     Project project = new Project(workspaceId, dbVendorId, name, description);
