+++ apps/backend/core/src/test/java/com/schemafy/core/project/application/service/CreateProjectServiceTest.java
@@ -35,7 +35,7 @@ import static org.mockito.BDDMockito.then;
 @DisplayName("CreateProjectService")
 class CreateProjectServiceTest {
 
-  private static final Long REQUESTED_DB_VENDOR_ID = 999L;
+  private static final Integer REQUESTED_DB_VENDOR_ID = 999;
   private static final String WORKSPACE_ID = "workspace-id";
   private static final String REQUESTER_ID = "requester-id";
